Visualize a causal forest results object

Certain plots only require a results object and not the other components of a cf_eval object. These can be created from results only.

# S3 method for results
plot(x, kind, ...)

Arguments

x
A results object

kind
The type of plot to create

...
Additional arguments passed to subsequent plot functions

Value

A plot

Details

Possible options for kind are:

cate

A density plot of estimated conditional average treatment effects, i.e. the causal forest predictions. The most straightforward way to look for treatment effect heterogeneity.

bias

A histogram of each observation's contribution to the overall bias of the model, relative to a simple difference in means.

propensities

A histogram of fitted propensities. The causal forest requires the assumption that we cannot deterministically tell the treatment status of an individual given its covariates. In other words, none of the propensity scores should be near zero or one.

See also

Examples

if (FALSE) {
 require(grf)

 Xdat <- subset(cfex, select = -c(W, Y))
 X <- make_contrasts(Xdat, 'fct')
 cf <- causal_forest(X, cfex$Y, cfex$W)

 cfe <- cf_eval(cf, Xdat)
 plot(cfe$res, kind = 'bias')
}