Assess counterfactual performance of a model capable of predictions under interventions

Assess counterfactual performance of a model capable of predictions under interventions

Usage

CFscore(
  data,
  model,
  predictions,
  Y,
  propensity_formula,
  ip,
  A,
  treatments,
  bootstrap,
  plot = TRUE,
  quiet = FALSE
)

Arguments

data

A data.frame on which the model is to be validated.

model

A glm (lm?) which can make predictions under interventions which has treatment as a predictor, or a list of glm's for each specified treatment in the treatments variable. Either model or predictions must be given.

predictions

A numeric vector of predictions under intervention, or a list of numeric vectors of predictions under different interventions as specified by treatments variable. Either model or predictions must be given.

Y

A character string indicating the name of the observed outcome column in data, or a numeric vector of the observed outcomes.

propensity_formula

A formula used to estimate the inverse-probability weights for the validation data. Treatment variable should be on the left hand side, all confounders on the right hand side. It is possible that there is a different set of confounders in the validation dataset compared to the model-development dataset. Either this or ipw must be given

ip

The inverse probabilty weights for the validation data. Can be either string indicating the name of the ip column in the validation data, or a numeric vector of ip-weights.

A

A character string indicating the name of the realized treatment column in data. Must be given if ipw's are specified. It is automatically inferred from propensity_formula if given.

treatments

A treatment level or a list of all treatment levels for which the counterfactual perormance measures should be evaluated.

bootstrap

If a 95 bootstrappingis desired, this can be achieved by setting this variable to a integer indicating the amount of iterations. Argument propensity_formula must then be given.

plot

If set to TRUE, generate calibration plot

quiet

If set to TRUE, don't print assumptions

Value

A list of performance measures (Brier, Observed/Expected, AUC) of the model on counterfactual data, where each subject is assigned to the treatments as given.

Examples

causal_model <- glm(
  Y ~ A + P,
  family = "binomial",
  data = df_dev,
  weights = ip_weights(df_dev, A ~ L)
)
#> Warning: non-integer #successes in a binomial glm!
CFscore(data = df_val, model = causal_model, Y = "Y",
        propensity_formula = A ~ L, treatments = list(0,1))
#> Estimation of the performance of the prediction model in a
#>  counterfactual (CF) dataset where everyone's treatment was set to 0 and
#>  1.
#> The following assumptions must be satisfied for correct inference:
#> - Conditional exchangeability requires that {L} are sufficient to
#>  adjust for confounding and selection bias between treatment and
#>  outcome.
#> - Positivity (assess $weights for outliers)
#> - Consistency
#> - No interference
#> - No measurement error
#> - Correctly specified propensity formula
#>    metric   CF0   CF1
#> 1   Brier 0.197 0.212
#> 2     auc 0.753 0.726
#> 3 OEratio 1.003 0.991